requirements.txt
matplotlib.pyplot
matplotlib.patches
streamlit